Add the admin pages
Set lockout in Startup class. Add SentAt to Report model. Delele
Blocked property (use inherited LockoutEnabled and LockoutEnd) from
User. Add ReportsController.GetReports() unit test and update the
method to pass it. Add report service to the Angular project.
Update the report component. Create admin component and map it
to '/Admin' path in the router. Add Lockout controller and use it
in the lockout method of the user service. Add block component
and map it to '/Block/:id' path in the router.